requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
django-rest-auth
django-allauth
django-cors-headers
djangorestframework-jwt
Pillow
drf-spectacular